# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "MCS_Server API 接口文档",
        "title": "Swagger MCS_Server API",
        "contact": {},
        "version": "0.0.1"
    },
    "basePath": "/",
    "paths": {
        "/consult/add": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "问诊"
                ],
                "summary": "创建一个问诊信息",
                "parameters": [
                    {
                        "description": "问诊信息的各个字段",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AddConsult"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"问诊信息添加成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/consult/findByUser": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "问诊"
                ],
                "summary": "根据名称用户id查询该用户的问诊信息",
                "parameters": [
                    {
                        "description": "用户id",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.FindConsultByUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"问诊信息获取成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/prescription/add": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "处方"
                ],
                "summary": "增加一个处方",
                "parameters": [
                    {
                        "description": "机构id，用户id，问诊信息id，处方类型，医生id，医生名称",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AddPrescription"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"处方新增成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "request.AddConsult": {
            "type": "object",
            "properties": {
                "consult_status": {
                    "type": "integer"
                },
                "create_user_id": {
                    "type": "string"
                },
                "dept_id": {
                    "type": "integer"
                },
                "dept_name": {
                    "type": "string"
                },
                "diagnosis": {
                    "type": "string"
                },
                "doctor_id": {
                    "type": "string"
                },
                "doctor_level_code": {
                    "type": "string"
                },
                "doctor_level_name": {
                    "type": "string"
                },
                "doctor_name": {
                    "type": "string"
                },
                "drug_ids": {
                    "type": "string"
                },
                "drug_names": {
                    "type": "string"
                },
                "org_id": {
                    "type": "integer"
                },
                "org_name": {
                    "type": "string"
                },
                "person_age": {
                    "type": "integer"
                },
                "person_birth_date": {
                    "type": "string"
                },
                "person_card_id": {
                    "type": "string"
                },
                "person_card_type": {
                    "type": "string"
                },
                "person_gender_code": {
                    "type": "string"
                },
                "person_gender_name": {
                    "type": "string"
                },
                "person_name": {
                    "type": "string"
                },
                "person_phone_no": {
                    "type": "string"
                },
                "photo_ids": {
                    "type": "string"
                },
                "question": {
                    "type": "string"
                }
            }
        },
...
        "request.AddPrescription": {
            "type": "object",
            "properties": {
                "consult_id": {
                    "type": "integer"
                },
                "doctor_id": {
                    "type": "string"
                },
                "doctor_name": {
                    "type": "string"
                },
                "org_id": {
                    "type": "integer"
                },
                "prescription_type": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "request.FindConsultByUser": {
            "type": "object",
            "properties": {
                "create_user_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "x-token",
            "in": "header"
        }
    }
}